Scrabble/classes (#12)
* adding new trait and initial class work

* updating stat tooltips to respect class changes

* adding wizardBlock for wizard attacks

* merge fix

* change max traits to 3

* wizard select and combat

* getting ready for beta

* fixing import break

* fixing rebirth stats

* many fixes for beta branch

* nit fixes, hopefully preventing random breaks in future

* adding more hotkeys

* ui fixes

* nit

+++ src/scenes/RegionScene.js
@@ -49,6 +49,11 @@ export class RegionScene extends SceneUIBase {
         this.fiberKey = undefined;
         this.stoneKey = undefined;
         this.crystalKey = undefined;
+        this.marketKey = undefined;
+        this.tavernKey = undefined;
+        this.warehouseKey = undefined;
+        this.alchemyKey = undefined;
+        this.destroyKey = undefined;
         this.hoveredTile = [-1, -1];
         this.regionTiles = [];
         this.regionStats = undefined;
@@ -293,7 +298,7 @@ export class RegionScene extends SceneUIBase {
                 this._updateRegionStats();
                 break;
             case "destroy":
-                if (PlayerData.instance.gold >= blob.tile.building.tier * Statics.DESTROY_BUILDING_COST) {
+                if (blob.tile.building.name !== "Town" && PlayerData.instance.gold >= blob.tile.building.tier * Statics.DESTROY_BUILDING_COST) {
                     PlayerData.instance.addGold(-blob.tile.building.tier * Statics.DESTROY_BUILDING_COST);
                     this.region.destroyBuilding(blob.tile.x, blob.tile.y);
                 }
@@ -444,6 +449,7 @@ export class RegionScene extends SceneUIBase {
 
     refresh() {
         this.rebirth();
+        this._onRegionChanged();
     }
 
     changeRegion() {
@@ -531,16 +537,18 @@ export class RegionScene extends SceneUIBase {
 
         this.invasionLabel = this.add.bitmapText(this.relativeX(720), this.relativeY(90), "courier20", "Invasion", 20, 1);
         this.invasionLabel.setVisible(this.progression.unlocks.buildings);
-        var invasionPercent = this.region.invasionCounter / Statics.INVASION_THRESHOLD;
         this.invasionLabel.setTint(Phaser.Display.Color.GetColor(255, 255, 255));
 
         this.offlineLabel = this.add.bitmapText(this.relativeX(660), this.relativeY(140), "courier20", "Offline Time: " + WorldData.instance.time.getOfflineTimeString());
+        this.speed0xButton = new TextButton(this, this.relativeX(760), this.relativeY(160), 30, 20, "0x")
+            .onClickHandler(() => { this._setTimeScale(0); });
         this.speed1xButton = new TextButton(this, this.relativeX(795), this.relativeY(160), 30, 20, "1x")
-            .onClickHandler(() => { WorldData.instance.time.setTimeScale(1); });
+            .onClickHandler(() => { this._setTimeScale(1); });
+        this.speed1xButton.setTextColor(Phaser.Display.Color.GetColor(255, 255, 0));
         this.speed2xButton = new TextButton(this, this.relativeX(830), this.relativeY(160), 30, 20, "2x")
-            .onClickHandler(() => { WorldData.instance.time.setTimeScale(2); });
+            .onClickHandler(() => { this._setTimeScale(2); });
         this.speed5xButton = new TextButton(this, this.relativeX(865), this.relativeY(160), 30, 20, "5x")
-            .onClickHandler(() => { WorldData.instance.time.setTimeScale(5); });
+            .onClickHandler(() => { this._setTimeScale(5); });
 
         this.autoExploreLabel = this.add.bitmapText(this.relativeX(660), this.relativeY(190), "courier20", "Auto Explore:");
         this.autoExploreButton = new TextButton(this, this.relativeX(795), this.relativeY(190), 40, 20, "OFF")
@@ -591,6 +599,11 @@ export class RegionScene extends SceneUIBase {
         this.roadKey = this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.R);
         this.productionKey = this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.P);
         this.exploreKey = this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.E);
+        this.marketKey = this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.M);
+        this.tavernKey = this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.T);
+        this.warehouseKey = this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.W);
+        this.alchemyKey = this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.A);
+        this.destroyKey = this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.D);
         this.woodKey = this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.ONE);
         this.leatherKey = this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.TWO);
         this.metalKey = this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.THREE);
@@ -599,6 +612,28 @@ export class RegionScene extends SceneUIBase {
         this.crystalKey = this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.SIX);
         this._updateRegionStats();
         this._onRegionChanged();
+    }
+
+    _setTimeScale(value) {
+        WorldData.instance.time.setTimeScale(value);
+        this.speed0xButton.setTextColor(Phaser.Display.Color.GetColor(255, 255, 255));
+        this.speed1xButton.setTextColor(Phaser.Display.Color.GetColor(255, 255, 255));
+        this.speed2xButton.setTextColor(Phaser.Display.Color.GetColor(255, 255, 255));
+        this.speed5xButton.setTextColor(Phaser.Display.Color.GetColor(255, 255, 255));
+        switch (value) {
+            case 0:
+                this.speed0xButton.setTextColor(Phaser.Display.Color.GetColor(255, 255, 0));
+                break;
+            case 1:
+                this.speed1xButton.setTextColor(Phaser.Display.Color.GetColor(255, 255, 0));
+                break;
+            case 2:
+                this.speed2xButton.setTextColor(Phaser.Display.Color.GetColor(255, 255, 0));
+                break;
+            case 5:
+                this.speed5xButton.setTextColor(Phaser.Display.Color.GetColor(255, 255, 0));
+                break;
+        }
     }
 
     update(__time, delta) {
@@ -665,6 +700,30 @@ export class RegionScene extends SceneUIBase {
             this._tileActionHandler("build", {
                 tile: this.region.map[this.hoveredTile[1]][this.hoveredTile[0]],
                 building: BuildingRegistry.getBuildingByName('crystal')
+            });
+        } else if (Phaser.Input.Keyboard.JustUp(this.marketKey) && this.hoveredTile[0] !== -1) {
+            this._tileActionHandler("build", {
+                tile: this.region.map[this.hoveredTile[1]][this.hoveredTile[0]],
+                building: BuildingRegistry.getBuildingByName('market')
+            });
+        } else if (Phaser.Input.Keyboard.JustUp(this.tavernKey) && this.hoveredTile[0] !== -1) {
+            this._tileActionHandler("build", {
+                tile: this.region.map[this.hoveredTile[1]][this.hoveredTile[0]],
+                building: BuildingRegistry.getBuildingByName('tavern')
+            });
+        } else if (Phaser.Input.Keyboard.JustUp(this.warehouseKey) && this.hoveredTile[0] !== -1) {
+            this._tileActionHandler("build", {
+                tile: this.region.map[this.hoveredTile[1]][this.hoveredTile[0]],
+                building: BuildingRegistry.getBuildingByName('warehouse')
+            });
+        } else if (Phaser.Input.Keyboard.JustUp(this.alchemyKey) && this.hoveredTile[0] !== -1) {
+            this._tileActionHandler("build", {
+                tile: this.region.map[this.hoveredTile[1]][this.hoveredTile[0]],
+                building: BuildingRegistry.getBuildingByName('alchemy')
+            });
+        } else if (Phaser.Input.Keyboard.JustUp(this.destroyKey) && this.hoveredTile[0] !== -1) {
+            this._tileActionHandler("destroy", {
+                tile: this.region.map[this.hoveredTile[1]][this.hoveredTile[0]]
             });
         }
